requirements.txt
alembic==1.7.7
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coverage==6.4
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
grpcio==1.44.0
grpcio-tools==1.44.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
numpy==1.22.4
pandas==1.4.2
protobuf==3.20.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.36
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
